requirements.txt
PyQt6
pynput
sounddevice
soundfile
numpy
mlx-whisper
ollama
pyperclip
pypinyin
opencc-python-reimplemented
pyobjc-framework-ApplicationServices
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
